*** empty log message ***

+++ Tools.cpp
@@ -2,7 +2,7 @@
  * @author Sylvain Filteau <admin@cidsphere.com>
  * @author Philippe Tremblay <pht@infodev.ca>
  * @date avril 2004
- * @version 0.5
+ * @version 0.6
  *
  * \brief Corp de la classe CTools
  */
@@ -29,10 +29,10 @@ int CTools::MsToNbSamples(int ms, int debit)
 
 int CTools::NbEchantillonsParCycle(float vitesse, float frequence)
 {
-	return (frequence/((1/vitesse)*1000));
+	return (int)(frequence/((1/vitesse)*1000));
 }
 
 short CTools::Mixer(short son1, short son2, float mix)
 {
-	return ((1-mix) * son1) + (mix * son2);
+	return (short)((1-mix) * son1) + (mix * son2);
 }
